backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.5.2
pandas==2.1.4
numpy==1.26.2
nltk==3.8.1
newsapi-python==0.2.7
joblib==1.3.2
textblob==0.17.1
requests==2.31.0